GroupBy output type.
