+++ query-engine/core/src/schema_builder/output_types/aggregation/group_by.rs
@@ -6,75 +6,84 @@ use super::*;
 /// Builds group by aggregation object type for given model (e.g. GroupByUserOutputType).
 pub(crate) fn group_by_output_object_type(ctx: &mut BuilderContext, model: &ModelRef) -> ObjectTypeWeakRef {
     let ident = Identifier::new(
-        format!("GroupBy{}OutputType", capitalize(&model.name)),
+        format!("{}GroupByOutputType", capitalize(&model.name)),
         PRISMA_NAMESPACE,
     );
     return_cached_output!(ctx, &ident);
 
     let object = ObjectTypeStrongRef::new(ObjectType::new(ident.clone(), Some(ModelRef::clone(model))));
-    let mut fields = vec![count_field()];
 
+    // Model fields that can be grouped by value.
+    let mut object_fields = scalar_fields(model);
+
+    // Fields used in aggregations
+    let non_list_fields = collect_non_list_fields(model);
+    let numeric_fields = collect_numeric_fields(model);
+
+    // Count is available on all fields.
     append_opt(
-        &mut fields,
-        numeric_aggregation_field(ctx, "avg", &model, field_avg_output_type),
+        &mut object_fields,
+        aggregation_field(ctx, "count", &model, model.fields().scalar(), field_avg_output_type),
     );
 
     append_opt(
-        &mut fields,
-        numeric_aggregation_field(ctx, "sum", &model, map_scalar_output_type),
+        &mut object_fields,
+        aggregation_field(ctx, "avg", &model, numeric_fields.clone(), field_avg_output_type),
     );
 
     append_opt(
-        &mut fields,
-        numeric_aggregation_field(ctx, "min", &model, map_scalar_output_type),
+        &mut object_fields,
+        aggregation_field(ctx, "sum", &model, numeric_fields.clone(), map_scalar_output_type),
     );
 
     append_opt(
-        &mut fields,
-        numeric_aggregation_field(ctx, "max", &model, map_scalar_output_type),
+        &mut object_fields,
+        aggregation_field(ctx, "min", &model, non_list_fields.clone(), map_scalar_output_type),
     );
 
-    object.set_fields(fields);
+    append_opt(
+        &mut object_fields,
+        aggregation_field(ctx, "max", &model, non_list_fields, map_scalar_output_type),
+    );
+
+    object.set_fields(object_fields);
     ctx.cache_output_type(ident, ObjectTypeStrongRef::clone(&object));
 
     ObjectTypeStrongRef::downgrade(&object)
 }
 
-pub(crate) fn count_field() -> OutputField {
-    field("count", vec![], OutputType::int(), None)
+fn scalar_fields(model: &ModelRef) -> Vec<OutputField> {
+    let fields = model.fields().scalar();
+
+    fields
+        .into_iter()
+        .map(|f| field(f.name.clone(), vec![], map_scalar_output_type(&f), None).optional())
+        .collect()
 }
 
-/// Returns an aggregation field with given name if the model contains any numeric fields.
-/// Fields inside the object type of the field may have a fixed output type.
-pub(crate) fn numeric_aggregation_field<F>(
+/// Returns an aggregation field with given name if the passed fields contains any fields.
+/// Field types inside the object type of the field are determined by the passed mapper fn.
+pub(crate) fn aggregation_field<F>(
     ctx: &mut BuilderContext,
     name: &str,
     model: &ModelRef,
+    fields: Vec<ScalarFieldRef>,
     type_mapper: F,
 ) -> Option<OutputField>
 where
     F: Fn(&ScalarFieldRef) -> OutputType,
 {
-    let numeric_fields = collect_numeric_fields(model);
-
-    if numeric_fields.is_empty() {
+    if fields.is_empty() {
         None
     } else {
-        let object_type = OutputType::object(map_numeric_field_aggregation_object(
-            ctx,
-            model,
-            name,
-            &numeric_fields,
-            type_mapper,
-        ));
+        let object_type = OutputType::object(map_field_aggregation_object(ctx, model, name, &fields, type_mapper));
 
         Some(field(name, vec![], object_type, None).optional())
     }
 }
 
-/// Maps the object type for aggregations that operate on a (numeric) field level, rather than the entire model.
-/// Fields inside the object may have a fixed output type.
-pub(crate) fn map_numeric_field_aggregation_object<F>(
+/// Maps the object type for aggregations that operate on a field level, rather than the entire model.
+pub(crate) fn map_field_aggregation_object<F>(
     ctx: &mut BuilderContext,
     model: &ModelRef,
     suffix: &str,
@@ -99,26 +108,4 @@ where
     ctx.cache_output_type(ident, object.clone());
 
     Arc::downgrade(&object)
-}
-
-fn field_avg_output_type(field: &ScalarFieldRef) -> OutputType {
-    match field.type_identifier {
-        TypeIdentifier::Int | TypeIdentifier::BigInt | TypeIdentifier::Float => OutputType::float(),
-        TypeIdentifier::Decimal => OutputType::decimal(),
-        _ => map_scalar_output_type(field),
-    }
-}
-
-fn collect_numeric_fields(model: &ModelRef) -> Vec<ScalarFieldRef> {
-    model
-        .fields()
-        .scalar()
-        .into_iter()
-        .filter(|f| {
-            matches!(
-                f.type_identifier,
-                TypeIdentifier::Int | TypeIdentifier::BigInt | TypeIdentifier::Float | TypeIdentifier::Decimal
-            )
-        })
-        .collect()
 }
