+++ query-engine/core/src/schema_builder/output_types/aggregation/plain.rs
@@ -1,6 +1,3 @@
-use prisma_models::ScalarFieldRef;
-
-use super::output_objects::map_scalar_output_type;
 use super::*;
 
 /// Builds plain aggregation object type for given model (e.g. AggregateUser).
@@ -91,32 +88,4 @@ where
     ctx.cache_output_type(ident, object.clone());
 
     Arc::downgrade(&object)
-}
-
-fn field_avg_output_type(field: &ScalarFieldRef) -> OutputType {
-    match field.type_identifier {
-        TypeIdentifier::Int | TypeIdentifier::BigInt | TypeIdentifier::Float => OutputType::float(),
-        TypeIdentifier::Decimal => OutputType::decimal(),
-        _ => map_scalar_output_type(field),
-    }
-}
-
-fn collect_non_list_fields(model: &ModelRef) -> Vec<ScalarFieldRef> {
-    model.fields().scalar().into_iter().filter(|f| !f.is_list).collect()
-}
-
-fn collect_numeric_fields(model: &ModelRef) -> Vec<ScalarFieldRef> {
-    model
-        .fields()
-        .scalar()
-        .into_iter()
-        .filter(|field| is_numeric(field))
-        .collect()
-}
-
-fn is_numeric(field: &ScalarFieldRef) -> bool {
-    matches!(
-        field.type_identifier,
-        TypeIdentifier::Int | TypeIdentifier::BigInt | TypeIdentifier::Float | TypeIdentifier::Decimal
-    )
 }
